+++ Models/DisplayGameInput.cs
@@ -6,10 +6,10 @@ namespace RotoMonsterUI
     {
         public string AwayTeamCode { get; set; }
         public string HomeTeamCode { get; set; }
-        public double AwayTeamProjectedRuns { get; set; }
-        public double HomeTeamProjectedRuns { get; set; }
-        public double AwayTeamCurrentRuns { get; set; }
-        public double HomeTeamCurrentRuns { get; set; }
+        public float AwayTeamProjectedRuns { get; set; }
+        public float HomeTeamProjectedRuns { get; set; }
+        public float AwayTeamCurrentRuns { get; set; }
+        public float HomeTeamCurrentRuns { get; set; }
         public string CurrentInning { get; set; }
         public bool IsGameFinished { get; set; }
         public bool IsGameLive { get; set; }
